refactor(skills): inline disk I/O into action handlers, delete disk.ts
importFromDisk and exportToDisk bodies now live directly in the
.withActions() handlers. importReferences and deriveReferenceId
move to private helpers in workspace.ts. SkillsWorkspaceClient
duck type and all its helper types are gone—the .withActions()
callback types the client directly.
